requirements.txt
numpy
tensorflow
pandas
matplotlib
seaborn
dvc
notebook
python-box==1.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
-e .